finished

+++ server/controller/OrdersController.js
@@ -1,11 +1,12 @@
 const Model = require('../db/models');
 const Utils = require('../../utils');
+const Sequelize = require('sequelize');
 
 const getAllOrders = (req, res) => {
   Model.Order.findAll({
     include: {
       model: Model.Product,
-      as: 'Product',
+      as: 'products',
       attributes: ['id', 'name', 'flavor', 'complement'],
       through: {
         model: Model.ProductsOrder,
@@ -18,20 +19,20 @@ const getAllOrders = (req, res) => {
       if (result.length === 0) {
         return res
           .status(404)
-          .json({ code: 404, message: "Don't have any user created yet!" });
+          .json({ code: 404, message: "Don't have any order created yet!" });
       }
 
       result = JSON.parse(JSON.stringify(result));
 
       const returnedOrders = result.map((order) => {
-        const productsList = order.Product.map((product) => ({
+        const productsList = order.products.map((product) => ({
           ...product,
           qtd: product.qtd.qtd,
         }));
 
         return {
           ...order,
-          Product: productsList,
+          products: productsList,
         };
       });
 
@@ -47,7 +48,15 @@ const getOrderById = (req, res) => {
 };
 
 const postOrder = (req, res) => {
-  Model.Order.create(req.body)
+  const { client_name, table, products } = req.body;
+  Utils.validationEmptyorNull(req.body.client_name, res, 'Client_name');
+  Utils.validationEmptyorNull(req.body.table, res, 'Table');
+  Utils.validationEmptyorNull(req.body.products, res, 'Products');
+  Model.Order.create({
+    client_name,
+    table,
+    products,
+  })
     .then((result) => {
       req.body.products.map((item) => {
         Model.Product.findOne({
@@ -92,21 +101,37 @@ const putOrder = (req, res) => {
     .then((result) => {
       if (result === null) {
         return res.status(404).json({ code: 404, message: 'Order not found!' });
+      } else if (result.status == 'pending') {
+        Model.Order.update(
+          { status, processedAt: Sequelize.literal('CURRENT_TIMESTAMP') },
+          {
+            where: {
+              id: req.params.orderid,
+            },
+          }
+        )
+          .then(() => {
+            Utils.getOrder(req, res, req.params.orderid);
+          })
+          .catch((err) => {
+            return res.status(400).json({ code: 400, message: err.message });
+          });
+      } else {
+        Model.Order.update(
+          { status },
+          {
+            where: {
+              id: req.params.orderid,
+            },
+          }
+        )
+          .then(() => {
+            Utils.getOrder(req, res, req.params.orderid);
+          })
+          .catch((err) => {
+            return res.status(400).json({ code: 400, message: err.message });
+          });
       }
-      Model.Order.update(
-        { status },
-        {
-          where: {
-            id: req.params.orderid,
-          },
-        }
-      )
-        .then(() => {
-          Utils.getOrder(req, res, req.params.orderid);
-        })
-        .catch((err) => {
-          return res.status(400).json({ code: 400, message: err.message });
-        });
     })
     .catch((err) => {
       return res.status(400).json({ code: 400, message: err.message });
@@ -114,24 +139,21 @@ const putOrder = (req, res) => {
 };
 
 const deleteOrder = (req, res) => {
-  Model.Order.destroy({
+  Model.Order.findOne({
     where: {
       id: req.params.orderid,
     },
   })
-    .then(() => {
-      res.status(200).send('Produto excluído com sucesso');
-    })
-    .catch((err) => {
-      return res.status(400).json({ code: 400, message: err.message });
-    });
-  Model.ProductsOrder.destroy({
-    where: {
-      order_id: req.params.orderid,
-    },
-  })
-    .then(() => {
-      res.status(200).send('Produto excluído com sucesso');
+    .then((result) => {
+      if (result !== null) {
+        Model.Order.destroy({
+          where: {
+            id: req.params.orderid,
+          },
+        });
+        return res.status(200).json(result);
+      }
+      return res.status(404).json({ code: 404, message: 'Order not found!' });
     })
     .catch((err) => {
       return res.status(400).json({ code: 400, message: err.message });
